// api/models/LikeRecordRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * 点赞请求参数
 */
export type LikeRecordRequest = {
    /**
     * 目标内容ID
     */
    targetId: number;
    /**
     * 目标类型(1-feed, 2-post, 3-module等)
     */
    targetType: string;
};